refactor: merge eth-node ens into protocol ens
